LayoutTests/fast/scrolling/latching/latched-scroll-in-passive-region-expected.txt

Tests that a gesture in the passive wheel event region dispatches more than one wheel event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS wheelEventHandlerCount > 4 is true
PASS successfullyParsed is true

TEST COMPLETE
